refactor(OpsScrapUserClient) : Ops 스크랩 유저 클라이언트 패키지 이동 및 API 경로 수정
- `OpsScrapUserClient` 클래스를 새로운 `scrap.config` 패키지로 이동하여 구성
- `ops.server.url` 값을 `@Value` 어노테이션을 사용해 프로퍼티 파일에서 읽어오도록 변경
- 스크랩 유저 조회 및 동기화 결과 알림 전송을 처리하는 메서드 `fetchUnsyncedUsers` 및 `notifyScrapResult`를 구현하여 OPS 서버와 통신하는 기능 추가
